requirements.txt
asgiref==3.2.10
astroid==2.4.2
Django==3.1.2
django-extensions==3.0.9
django-phonenumber-field==5.0.0
isort==5.6.4
lazy-object-proxy==1.4.3
markdown2==2.3.10
mccabe==0.6.1
phonenumbers==8.12.11
psycopg2==2.8.6
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
tzlocal==2.1
wrapt==1.12.1
